% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angiosperm_phylogeny}
\alias{angiosperm_phylogeny}
\title{Microsynteny-based angiosperm phylogeny.}
\format{
An object of class 'phylo'.
}
\usage{
data(angiosperm_phylogeny)
}
\description{
The data are stored as a 'phylo' object, which can be created
with \code{treeio::read.tree()}.
}
\examples{
data(angiosperm_phylogeny)
}
\references{
Zhao, T., Zwaenepoel, A., Xue, J. Y., Kao, S. M., Li, Z.,
Schranz, M. E., & Van de Peer, Y. (2021).
Whole-genome microsynteny-based phylogeny of angiosperms.
Nature Communications, 12(1), 1-14.
}
\keyword{datasets}
