% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_microsynteny-based_phylogeny.R
\name{binarize_and_transpose}
\alias{binarize_and_transpose}
\title{Binarize and transpose the phylogenomic profile matrix}
\usage{
binarize_and_transpose(profile_matrix = NULL)
}
\arguments{
\item{profile_matrix}{A matrix with phylogenomic profiles obtained
with \code{phylogenomic_profile}.}
}
\value{
A binary and transposed version of the profiles matrix.
}
\description{
Binarize and transpose the phylogenomic profile matrix
}
\examples{
data(clusters)
profile_matrix <- phylogenomic_profile(clusters)
tmat <- binarize_and_transpose(profile_matrix)
}
