% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blast_list}
\alias{blast_list}
\title{List of data frames containing BLAST-like tabular output}
\format{
A list of data frames containing the pairwise comparisons between
proteomes.
}
\usage{
data(blast_list)
}
\description{
The object was created using the example code, exactly as described
in the vignette.
}
\examples{
data(blast_list)
}
\keyword{datasets}
