% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_seqnames}
\alias{clean_seqnames}
\title{Clean sequence names}
\usage{
clean_seqnames(seq = NULL)
}
\arguments{
\item{seq}{A list of AAStringSet objects.}
}
\value{
A list of AAStringSet objects with clean names.
}
\description{
This function removes anything after a space in sequence names and adds
a unique species abbreviation to each sequence name.
}
\examples{
data(proteomes)
seq <- proteomes
clean_seqs <- clean_seqnames(seq)
}
