% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{Synteny network clusters of BUSCO genes for 25 eudicot species}
\format{
A 2-column data frame containing the following variables:
\describe{
\item{Gene}{Gene ID}
\item{Cluster}{Cluster ID}
}
}
\usage{
data(clusters)
}
\description{
Synteny network clusters of BUSCO genes for 25 eudicot species
}
\examples{
data(clusters)
}
\references{
Zhao, T., & Schranz, M. E. (2019). Network-based microsynteny analysis
identifies major differences and genomic outliers in mammalian and
angiosperm genomes. Proceedings of the National Academy of
Sciences, 116(6), 2165-2174.
}
\keyword{datasets}
