% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diamond_is_installed}
\alias{diamond_is_installed}
\title{Check if DIAMOND is installed}
\usage{
diamond_is_installed()
}
\value{
Logical indicating whether DIAMOND is installed or not.
}
\description{
Check if DIAMOND is installed
}
\examples{
diamond_is_installed()
}
