% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edges}
\alias{edges}
\title{Data frame containing a network represented as an edge list}
\format{
A data frame containing anchor between two
proteomes.
}
\usage{
data(edges)
}
\description{
The object was created using the example code, exactly as described
in the vignette.
}
\examples{
data(edges)
}
\keyword{datasets}
