% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_import.R
\name{fasta2AAStringSetlist}
\alias{fasta2AAStringSetlist}
\title{Read FASTA files in a directory as a list of AAStringSet objects}
\usage{
fasta2AAStringSetlist(fasta_dir)
}
\arguments{
\item{fasta_dir}{Character indicating the path to the directory containing
FASTA files.}
}
\value{
A list of AAStringSet objects, where each element represents a
different FASTA file.
}
\description{
Read FASTA files in a directory as a list of AAStringSet objects
}
\examples{
fasta_dir <- system.file("extdata", "sequences", package = "syntenet")
aastringsetlist <- fasta2AAStringSetlist(fasta_dir)
}
