% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_import.R
\name{gff2GRangesList}
\alias{gff2GRangesList}
\title{Read GFF/GTF files in a directory as a GRangesList object}
\usage{
gff2GRangesList(gff_dir)
}
\arguments{
\item{gff_dir}{Character indicating the path to the directory containing
GFF/GTF files.}
}
\value{
A GRangesList object, where each element represents a different
GFF/GTF file.
}
\description{
Read GFF/GTF files in a directory as a GRangesList object
}
\examples{
gff_dir <- system.file("extdata", "annotation", package = "syntenet")
grangeslist <- gff2GRangesList(gff_dir)
}
