% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_synteny.R
\name{intraspecies_synteny}
\alias{intraspecies_synteny}
\title{Detect intraspecies synteny}
\usage{
intraspecies_synteny(
  blast_intra = NULL,
  intra_dir = file.path(tempdir(), "intra"),
  annot_list = NULL,
  anchors = 5,
  max_gaps = 25,
  is_pairwise = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{blast_intra}{A list of BLAST data frames for intraspecies comparisons.}

\item{intra_dir}{Output directory.}

\item{annot_list}{A processed GRangesList or CompressedGRangesList object
as returned by \code{process_input()}.}

\item{anchors}{Numeric indicating the minimum required number of genes
to call a syntenic block. Default: 5.}

\item{max_gaps}{Numeric indicating the number of upstream and downstream
genes to search for anchors. Default: 25.}

\item{is_pairwise}{Logical indicating if only pairwise blocks should be
reported. Default: TRUE.}

\item{verbose}{Logical indicating if log messages should be printed on
screen. Default: FALSE.}

\item{...}{Any additional arguments to
\code{mcscanx}.}
}
\value{
Paths to .collinearity files.
}
\description{
Detect intraspecies synteny
}
\examples{
data(blast_list)
data(annotation)
blast_intra <- blast_list[1]
annot <- as.data.frame(annotation[[1]])
annot <- annot[, c("seqnames", "gene_id", "start", "end")]
annot_list <- list(Olucimarinus = annot)
intrasyn <- intraspecies_synteny(blast_intra, annot_list = annot_list)
}
