% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iqtree_is_installed}
\alias{iqtree_is_installed}
\title{Check if IQTREE is installed}
\usage{
iqtree_is_installed()
}
\value{
Logical indicating whether IQTREE is installed or not.
}
\description{
Check if IQTREE is installed
}
\examples{
iqtree_is_installed()
}
