% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{iqtree_version}
\alias{iqtree_version}
\title{Get IQ-TREE version}
\usage{
iqtree_version()
}
\value{
Numeric indicating IQ-TREE version, with either 1 or 2.
}
\description{
Get IQ-TREE version
}
\examples{
iqtree_version()
}
