% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_is_installed}
\alias{last_is_installed}
\title{Check if last is installed}
\usage{
last_is_installed()
}
\value{
Logical indicating whether last is installed or not.
}
\description{
Check if last is installed
}
\examples{
last_is_installed()
}
