% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Synteny network for the legume species G. max, P. vulgaris, and C. cajan}
\format{
An edgelist (i.e., a 2-column data frame with node 1 in
column 1 and node 2 in column 2).
}
\usage{
data(network)
}
\description{
Synteny network for the legume species G. max, P. vulgaris, and C. cajan
}
\examples{
data(network)
}
\references{
Zhao, T., & Schranz, M. E. (2019). Network-based microsynteny analysis
identifies major differences and genomic outliers in mammalian and
angiosperm genomes. Proceedings of the National Academy of
Sciences, 116(6), 2165-2174.
}
\keyword{datasets}
