% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_phylogenomic_profiling.R
\name{phylogenomic_profile}
\alias{phylogenomic_profile}
\title{Perform phylogenomic profiling for synteny network clusters}
\usage{
phylogenomic_profile(clusters = NULL)
}
\arguments{
\item{clusters}{A 2-column data frame with variables \strong{Gene} and
\strong{Cluster} as returned by \code{cluster_network}.}
}
\value{
A matrix of i rows and j columns containing
the number of genes in cluster i for each species j. The number of rows
is equal to the number of clusters in \strong{clusters}, and
the number of columns is equal to the number of species
in \strong{clusters}.
}
\description{
Perform phylogenomic profiling for synteny network clusters
}
\examples{
data(clusters)
profiles <- phylogenomic_profile(clusters)
}
