% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot a heatmap of phylogenomic profiles}
\usage{
plot_profiles(
  profiles = NULL,
  species_annotation = NULL,
  palette = "Greens",
  cluster_species = FALSE,
  cluster_columns = profiles$hclust,
  show_colnames = FALSE,
  discretize = TRUE,
  ...
)
}
\arguments{
\item{profiles}{A list of phylogenomic profiles as returned
by \code{phylogenomic_profile()}.}

\item{species_annotation}{A 2-column data frame with species IDs in
the first column (same as column names of profile matrix), and species
annotation (e.g., higher-level taxonomic information) in the second column.}

\item{palette}{A character vector of colors or a character scalar with
the name of an RColorBrewer palette. Default: "RdYlBu".}

\item{cluster_species}{Either a logical scalar (TRUE or FALSE) or
a character vector with the order of columns. Ideally, the character vector
should contain the order of species in a phylogenetically meaningful way.
If users have a species tree, they can read it
with \code{treeio::read.tree()} and get the species order
with \code{ggtree::get_taxa_name()}.}

\item{cluster_columns}{Either a logical scalar (TRUE or FALSE) or an hclust
object. Default: hclust object in the \strong{profiles} list as returned
by \code{phylogenomic_profile()}.}

\item{show_colnames}{Logical indicating whether to show column names (i.e.,
cluster IDs) or not. Showing cluster IDs can be useful when visualizing
a small subset of them. When visualizing all clusters, cluster IDs are
impossible to read. Default: FALSE.}

\item{discretize}{Logical indicating whether to discretize clusters in
4 categories: 0, 1, 2, and 3+.}

\item{...}{Additional parameters to \code{pheatmap::pheatmap()}.}
}
\value{
A pheatmap object.
}
\description{
Plot a heatmap of phylogenomic profiles
}
\examples{
data(clusters)
profiles <- phylogenomic_profile(clusters)
species_order <- c(
    "vra", "van", "pvu", "gma", "cca", "tpr", "mtr", "adu", "lja",
    "Lang", "car", "pmu", "ppe", "pbr", "mdo", "roc", "fve",
    "Mnot", "Zjuj", "jcu", "mes", "rco", "lus", "ptr"
) 
species_annotation <- data.frame(
   Species = species_order,
   Family = c(rep("Fabaceae", 11), rep("Rosaceae", 6),
              "Moraceae", "Ramnaceae", rep("Euphorbiaceae", 3), 
              "Linaceae", "Salicaceae")
)
# Option 1: cluster cols using ward.D clustering based on Jaccard distances
p <- plot_profiles(profiles, species_annotation, 
                   cluster_species = species_order)
                   
# Option 2: hierarchical clustering
p <- plot_profiles(profiles, species_annotation, 
                   cluster_species = species_order, 
                   cluster_columns = TRUE)
}
