% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{process_input}
\alias{process_input}
\title{Process sequence data}
\usage{
process_input(seq = NULL)
}
\arguments{
\item{seq}{A list of AAStringSet objects.}
}
\description{
Process sequence data
}
\examples{
data(proteomes)
}
