% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_similarity_searches.R
\name{read_last}
\alias{read_last}
\title{Read last/BLAST tables as a list of data frames}
\usage{
read_last(last_dir = NULL)
}
\arguments{
\item{last_dir}{Path to directory containing the tabular output
of last or similar programs (e.g., BLAST).}
}
\value{
A list of data frames with the tabular last output.
}
\description{
Read last/BLAST tables as a list of data frames
}
\examples{
# Path to output directory
last_dir <- system.file("extdata", package = "syntenet")

# Read output
l <- read_last(last_dir)
}
