% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scerevisiae_annot}
\alias{scerevisiae_annot}
\title{Genome annotation of the yeast species S. cerevisiae}
\format{
A GRangesList as returned by \code{process_input()} containing
the element \strong{Scerevisiae}.
}
\usage{
data(scerevisiae_annot)
}
\description{
Data obtained from Ensembl Fungi. Only annotation data for primary
transcripts were included.
}
\examples{
data(scerevisiae_annot)
}
\keyword{datasets}
