% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scerevisiae_diamond}
\alias{scerevisiae_diamond}
\title{Intraspecies DIAMOND output for S. cerevisiae}
\format{
A list of data frames (length 1) containing the whole paranome of
S. cerevisiae resulting from intragenome similarity searches.
}
\usage{
data(scerevisiae_diamond)
}
\description{
List obtained with \code{run_diamond()}.
}
\examples{
data(scerevisiae_diamond)
}
\keyword{datasets}
