% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenomic_profiling.R
\name{cluster_network}
\alias{cluster_network}
\title{Cluster the synteny network using the Infomap algorithm}
\usage{
cluster_network(network = NULL)
}
\arguments{
\item{network}{A network represented as an edge list, which is a 2-column
data frame with node 1 in the first column and node 2 in the second column.
In a synteny network, node 1 and node are the anchor pairs.}
}
\value{
A 2-column data frame with the following variables:
\describe{
\item{Gene}{Gene ID.}
\item{Cluster}{Cluster ID as identified by infomap.}
}
}
\description{
Cluster the synteny network using the Infomap algorithm
}
\examples{
library(cogeqc)
data(synnet)
network <- synnet[1:500, ]
clusters <- cluster_network(network)
}
