% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edges}
\alias{edges}
\title{Synteny network of Ostreococcus genomes represented as an edge list}
\format{
A data frame containing anchor pairs between two Ostreococcus
proteomes.
}
\usage{
data(edges)
}
\description{
The object was created by running \code{infer_syntenet} on the
\strong{blast_list} example data. Code to recreate this data set is
available at the script/ subdirectory.
}
\examples{
data(edges)
}
\keyword{datasets}
