% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny_detection.R
\name{parse_collinearity}
\alias{parse_collinearity}
\title{Parse .collinearity files into an edge list}
\usage{
parse_collinearity(collinearity_paths = NULL)
}
\arguments{
\item{collinearity_paths}{Character vector of paths to .collinearity files.}
}
\value{
A 2-column data frame with each gene from the anchor pair in each
column.
}
\description{
Parse .collinearity files into an edge list
}
\examples{
collinearity_paths <- system.file("extdata", "Olu.collinearity", 
                                  package = "syntenet")
net <- parse_collinearity(collinearity_paths)
}
