% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proteomes}
\alias{proteomes}
\alias{annotation}
\title{Filtered proteomes of Ostreococcus sp. species}
\format{
A list of AAStringSet objects containing
the elements \code{Olucimarinus}, \code{Osp_RCC809}, and \code{Otauri}.

A CompressedGRangesList containing
the elements \code{Olucimarinus}, \code{Osp_RCC809}, and \code{Otauri}.
}
\usage{
data(proteomes)

data(annotation)
}
\description{
Data obtained from Pico-PLAZA 3.0. Only the translated sequences of primary
transcripts were included, and only genes from chromosomes 1, 2, and 3.

Data obtained from Pico-PLAZA 3.0. Only annotation data for primary
transcripts were included, and only genes for chromosomes 1, 2, and 3.
}
\examples{
data(proteomes)
data(annotation)
}
\references{
Van Bel, M., Silvestri, F., Weitz, E. M., Kreft, L., Botzki, A.,
Coppens, F., & Vandepoele, K. (2021). PLAZA 5.0: extending the scope
and power of comparative and functional genomics in plants.
Nucleic acids research.

Van Bel, M., Silvestri, F., Weitz, E. M., Kreft, L., Botzki, A.,
Coppens, F., & Vandepoele, K. (2021). PLAZA 5.0: extending the scope
and power of comparative and functional genomics in plants.
Nucleic acids research.
}
\keyword{datasets}
