% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_mcscanx_file}
\alias{rcpp_mcscanx_file}
\title{rcpp_mcscanx_file}
\usage{
rcpp_mcscanx_file(
  blast_file,
  gff_file,
  prefix = "out",
  outdir = "",
  match_score = 50L,
  gap_penalty = -1L,
  match_size = 5L,
  e_value = 1e-05,
  max_gaps = 25L,
  overlap_window = 5L,
  is_pairwise = FALSE,
  in_synteny = 0L,
  verbose = FALSE
)
}
\arguments{
\item{blast_file}{blast input}

\item{gff_file}{gff input}

\item{prefix}{output prefix (default: out)}

\item{outdir}{output directory (default: "")}

\item{match_score}{match score (default: 50)}

\item{gap_penalty}{gap penalty (default: -1)}

\item{match_size}{match_size (default: 5)}

\item{e_value}{e_value (default: 1e-5)}

\item{max_gaps}{max gaps (default: 25)}

\item{overlap_window}{overlap window (default: 5)}

\item{is_pairwise}{specify if only pairwise blocks should be reported
(default: FALSE)}

\item{in_synteny}{specify patterns of collinear blocks.
0: intra- and inter-species (default); 1: intra-species; 2: inter-species}

\item{verbose}{specify if verbose output (default: FALSE)}
}
\value{
list
}
\description{
MCSanX provides a clustering module for viewing the
relationship of collinear segments in multiple genomes (or heavily redundant
genomes). It takes the predicted pairwise segments from dynamic programming
(DAGchainer in particular) and then tries to build consensus segments from a
set of related, overlapping segments.
}
\references{
Wang et al. (2012) MCScanX: a toolkit for detection and
evolutionary analysis of gene synteny and collinearity.
\emph{Nucleic acids research}. \bold{40.7}, e49-e49.

Haas et al. (2004) DAGchainer: a tool for mining segmental
genome duplications and synteny. \emph{Bioinformatics}. \bold{20.18}
3643-3646.
}
\author{
Kristian K Ullrich
}
